requirements.txt
cog==0.6.1
colour-science==0.4.2
numpy==1.24.3
